+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased changes
+- Added support for Transfer To Encrypted wallet.
 
 ## 3.1.0
 
@@ -37,7 +38,6 @@
 - Added support for query `UnBanNode`
 - Added support for query `StartBaker`
 - Added support for query `StopBaker`
-- Added support for Transfer To Encrypted wallet.
 
 ## 3.0.0
 - Stronger typing for various places in the API.